% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack_ids.R
\name{unpack_ids.data.frame}
\alias{unpack_ids.data.frame}
\title{unpack_ids}
\usage{
\method{unpack_ids}{data.frame}(x, id_var = NULL, plural_to_singular = TRUE, verbose = FALSE)
}
\description{
unpack_ids
}
