% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack_ids.R
\name{unpack_ids}
\alias{unpack_ids}
\title{unpack_ids}
\usage{
unpack_ids(x, ...)
}
\description{
Unpack a vector of "packed IDs" (like facility IDs or category IDs)
}
\examples{
packed_ids <- c("c(1, 3)", "c(1:4)")
unpack_ids(packed_ids)

packed_data <- tibble(cat_ids = c("c(1, 3)", "c(1:4)"))
unpack_ids(packed_data, id_var = "cat_ids")
unpack_ids(packed_data)

}
