% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack_integers.R
\name{unpack_integers}
\alias{unpack_integers}
\title{Unpack a vector of "packed integers" (like facility IDs or category IDs)}
\usage{
unpack_integers(x, ...)
}
\description{
Unpack a vector of "packed integers" (like facility IDs or category IDs)
}
\examples{
packed_integers <- c("c(1, 3)", "c(1:4)")
unpack_integers(packed_integers)

packed_data <- tibble(cat_ids = c("c(1, 3)", "c(1:4)"))
unpack_integers(packed_data, var_name = "cat_ids")
unpack_integers(packed_data)

}
